+++ compiler-plugin/src/main/java/io/ballerina/stdlib/persist/compiler/Constants.java
@@ -1,7 +1,7 @@
 /*
- * Copyright (c) 2022, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ * Copyright (c) 2023, WSO2 LLC. (http://www.wso2.org) All Rights Reserved.
  *
- * WSO2 Inc. licenses this file to you under the Apache License,
+ * WSO2 LLC. licenses this file to you under the Apache License,
  * Version 2.0 (the "License"); you may not use this file except
  * in compliance with the License.
  * You may obtain a copy of the License at
@@ -18,84 +18,47 @@
 
 package io.ballerina.stdlib.persist.compiler;
 
-import io.ballerina.compiler.syntax.tree.LiteralValueToken;
-import io.ballerina.compiler.syntax.tree.NodeFactory;
-import io.ballerina.compiler.syntax.tree.SyntaxKind;
-import io.ballerina.compiler.syntax.tree.Token;
-
-import static io.ballerina.compiler.syntax.tree.AbstractNodeFactory.createEmptyMinutiaeList;
-
 /**
- * Constants for Persist compiler plugin.
+ * Constants class.
  */
-public class Constants {
+public final class Constants {
+    public static final String PERSIST_DIRECTORY = "persist";
+    public static final String TIME_MODULE = "time";
+    public static final String EMPTY_STRING = "";
+    public static final String ARRAY = "[]";
+    public static final String LS = System.lineSeparator();
 
-    public static final String TRUE = "true";
-    public static final String EMPTY = "";
-    public static final String FALSE = "false";
-    public static final String INSERT_METHOD_NAME = "runInsertQuery";
-    public static final String ENTITY = "persist:Entity";
-    public static final String UNNECESSARY_CHARS_REGEX = "\"|\\n";
-    public static final String TABLE_NAME = "tableName";
-    public static final String AUTO_INCREMENT = "persist:AutoIncrement";
-    public static final String INCREMENT = "increment";
-    public static final String RELATION = "persist:Relation";
-    public static final String KEY_COLUMNS = "keyColumns";
-    public static final String REFERENCE = "reference";
-    public static final String ONE = "1";
-    public static final String ON_DELETE = "onDelete";
-    public static final String ON_UPDATE = "onUpdate";
-    public static final String ONE_TO_ONE = "ONE";
-    public static final String ONE_TO_MANY = "MANY";
+    private Constants() {
+    }
 
-    public static final String READ_FUNCTION = "read";
-    public static final String EXECUTE_FUNCTION = "execute";
-    public static final String BACKTICK = "`";
-    public static final String SPACE = " ";
-    public static final String OPEN_BRACES = "( ";
-    public static final String CLOSE_BRACES = ") ";
-    public static final String ASCENDING = "ascending";
-    public static final String BAL_ESCAPE_TOKEN = "'";
-    public static final String PERSIST_INHERITANCE_NODE = "*persist:AbstractPersistClient;";
     /**
      * Constants related to Ballerina types.
      */
     public static final class BallerinaTypes {
+
         public static final String INT = "int";
         public static final String STRING = "string";
         public static final String BOOLEAN = "boolean";
         public static final String DECIMAL = "decimal";
         public static final String FLOAT = "float";
-        public static final String DATE = "time:Date";
-        public static final String TIME_OF_DAY = "time:TimeOfDay";
-        public static final String UTC = "time:Utc";
-        public static final String CIVIL = "time:Civil";
+        public static final String BYTE = "byte";
+
+        private BallerinaTypes() {
+        }
     }
 
     /**
-     * SQL keywords used to construct the query.
+     * Constants related to Ballerina time type.
      */
-    public static final class SQLKeyWords {
-        public static final String WHERE = "WHERE";
-        public static final String LIMIT = "LIMIT";
-        public static final String ORDERBY = "ORDER BY";
-        public static final String ORDER_BY_ASCENDING = "ASC";
-        public static final String ORDER_BY_DECENDING = "DESC";
-        public static final String NOT_EQUAL_TOKEN = "<>";
-        public static final String AND = "AND";
-        public static final String OR = "OR";
-    }
+    public static final class BallerinaTimeTypes {
 
-    /**
-     * Constant nodes used in code modification.
-     */
-    public static final class TokenNodes {
-        public static final Token INTERPOLATION_START_TOKEN = NodeFactory.createLiteralValueToken(
-                SyntaxKind.INTERPOLATION_START_TOKEN, "${", createEmptyMinutiaeList(), createEmptyMinutiaeList());
-        public static final Token INTERPOLATION_END_TOKEN = NodeFactory.createLiteralValueToken(
-                SyntaxKind.CLOSE_BRACE_TOKEN, "}", createEmptyMinutiaeList(), createEmptyMinutiaeList());
-        public static final LiteralValueToken BACKTICK_TOKEN = NodeFactory.createLiteralValueToken(
-                SyntaxKind.BACKTICK_TOKEN, BACKTICK, createEmptyMinutiaeList(), createEmptyMinutiaeList());
+        public static final String DATE = "Date";
+        public static final String TIME_OF_DAY = "TimeOfDay";
+        public static final String UTC = "Utc";
+        public static final String CIVIL = "Civil";
+
+        private BallerinaTimeTypes() {
+        }
     }
 
 }