+++ compiler-plugin/src/main/java/io/ballerina/stdlib/persist/compiler/Utils.java
@@ -1,49 +1,46 @@
 /*
- * Copyright (c) 2022, WSO2 Inc. (http:www.wso2.org) All Rights Reserved.
+ * Copyright (c) 2023, WSO2 LLC. (http://www.wso2.org) All Rights Reserved.
  *
- * WSO2 Inc. licenses this file to you under the Apache License,
+ * WSO2 LLC. licenses this file to you under the Apache License,
  * Version 2.0 (the "License"); you may not use this file except
  * in compliance with the License.
  * You may obtain a copy of the License at
  *
- *     http:www.apache.orglicensesLICENSE-2.0
+ * http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing,
  * software distributed under the License is distributed on an
  * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
+ * KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations
  * under the License.
  */
 
 package io.ballerina.stdlib.persist.compiler;
 
-import io.ballerina.compiler.syntax.tree.FromClauseNode;
-import io.ballerina.compiler.syntax.tree.LiteralValueToken;
-import io.ballerina.compiler.syntax.tree.NodeFactory;
-import io.ballerina.compiler.syntax.tree.RemoteMethodCallActionNode;
-import io.ballerina.compiler.syntax.tree.SyntaxKind;
+import io.ballerina.compiler.syntax.tree.Node;
 import io.ballerina.projects.plugins.SyntaxNodeAnalysisContext;
+import io.ballerina.projects.plugins.codeaction.CodeActionArgument;
+import io.ballerina.projects.plugins.codeaction.CodeActionExecutionContext;
 import io.ballerina.tools.diagnostics.Diagnostic;
-import io.ballerina.tools.diagnostics.DiagnosticFactory;
-import io.ballerina.tools.diagnostics.DiagnosticInfo;
+import io.ballerina.tools.diagnostics.DiagnosticProperty;
 import io.ballerina.tools.diagnostics.DiagnosticSeverity;
-import io.ballerina.tools.diagnostics.Location;
+import io.ballerina.tools.text.TextRange;
+import org.wso2.ballerinalang.compiler.diagnostic.properties.BNumericProperty;
+import org.wso2.ballerinalang.compiler.diagnostic.properties.BStringProperty;
 
-import static io.ballerina.compiler.syntax.tree.AbstractNodeFactory.createEmptyMinutiaeList;
-import static io.ballerina.stdlib.persist.compiler.Constants.READ_FUNCTION;
+import java.util.List;
 
 /**
- * This class includes utility functions.
+ * Class containing util functions.
  */
-public class Utils {
+public final class Utils {
 
-    protected static String eliminateDoubleQuotes(String text) {
-        return text.substring(1, text.length() - 1);
+    private Utils() {
     }
 
     public static boolean hasCompilationErrors(SyntaxNodeAnalysisContext context) {
-        for (Diagnostic diagnostic : context.semanticModel().diagnostics()) {
+        for (Diagnostic diagnostic : context.compilation().diagnosticResult().diagnostics()) {
             if (diagnostic.diagnosticInfo().severity() == DiagnosticSeverity.ERROR) {
                 return true;
             }
@@ -51,27 +48,90 @@ public class Utils {
         return false;
     }
 
-    public static void reportDiagnostic(SyntaxNodeAnalysisContext ctx, Location location, String code,
-                                        String message, DiagnosticSeverity diagnosticSeverity) {
-        DiagnosticInfo diagnosticInfo = new DiagnosticInfo(code, message, diagnosticSeverity);
-        ctx.reportDiagnostic(DiagnosticFactory.createDiagnostic(diagnosticInfo, location));
+    public static String stripEscapeCharacter(String name) {
+        return name.startsWith("'") ? name.substring(1) : name;
     }
 
-    public static LiteralValueToken getStringLiteralToken(String value) {
-        return NodeFactory.createLiteralValueToken(
-                SyntaxKind.STRING_LITERAL, value, createEmptyMinutiaeList(), createEmptyMinutiaeList());
-    }
-
-    public static boolean isQueryUsingPersistentClient(FromClauseNode fromClauseNode) {
-
-        // From clause should contain remote call invocation
-        if (fromClauseNode.expression() instanceof RemoteMethodCallActionNode) {
-            RemoteMethodCallActionNode remoteCall = (RemoteMethodCallActionNode) fromClauseNode.expression();
-            String functionName = remoteCall.methodName().name().text();
-
-            // Remote function name should be read
-            return functionName.trim().equals(READ_FUNCTION);
+    public static String getTypeName(Node processedTypeNode) {
+        String typeName = processedTypeNode.kind().stringValue();
+        switch (processedTypeNode.kind()) {
+            case HANDLE_TYPE_DESC:
+            case ANY_TYPE_DESC:
+            case ANYDATA_TYPE_DESC:
+            case NEVER_TYPE_DESC:
+                // here typename is not empty
+                break;
+            case UNION_TYPE_DESC:
+                typeName = "union";
+                break;
+            case NIL_TYPE_DESC:
+                typeName = "()";
+                break;
+            case MAP_TYPE_DESC:
+                typeName = "map";
+                break;
+            case ERROR_TYPE_DESC:
+                typeName = "error";
+                break;
+            case STREAM_TYPE_DESC:
+                typeName = "stream";
+                break;
+            case FUNCTION_TYPE_DESC:
+                typeName = "function";
+                break;
+            case TUPLE_TYPE_DESC:
+                typeName = "tuple";
+                break;
+            case TABLE_TYPE_DESC:
+                typeName = "table";
+                break;
+            case DISTINCT_TYPE_DESC:
+                typeName = "distinct";
+                break;
+            case INTERSECTION_TYPE_DESC:
+                typeName = "intersection";
+                break;
+            case FUTURE_TYPE_DESC:
+                typeName = "future";
+                break;
+            case RECORD_TYPE_DESC:
+                typeName = "in-line record";
+                break;
+            case OBJECT_TYPE_DESC:
+                typeName = "object";
+                break;
+            default:
+                if (typeName.isBlank()) {
+                    typeName = processedTypeNode.kind().name();
+                }
         }
-        return false;
+        return typeName;
     }
+
+    public static String getStringArgument(CodeActionExecutionContext context, String key) {
+        for (CodeActionArgument arg : context.arguments()) {
+            if (key.equals(arg.key())) {
+                return arg.valueAs(String.class);
+            }
+        }
+        return null;
+    }
+
+    public static TextRange getTextRangeArgument(CodeActionExecutionContext context, String key) {
+        for (CodeActionArgument arg : context.arguments()) {
+            if (key.equals(arg.key())) {
+                return arg.valueAs(TextRange.class);
+            }
+        }
+        return null;
+    }
+
+    public static String getStringDiagnosticProperty(List<DiagnosticProperty<?>> diagnosticProperties, int index) {
+        return ((BStringProperty) diagnosticProperties.get(index)).value();
+    }
+
+    public static int getNumericDiagnosticProperty(List<DiagnosticProperty<?>> diagnosticProperties, int index) {
+        return ((BNumericProperty) diagnosticProperties.get(index)).value().intValue();
+    }
+
 }