+++ compiler-plugin/src/main/java/module-info.java
@@ -21,5 +21,6 @@ module io.ballerina.stdlib.persist.plugin {
     requires io.ballerina.parser;
     requires io.ballerina.tools.api;
     exports io.ballerina.stdlib.persist.compiler;
-    exports io.ballerina.stdlib.persist.compiler.models;
+    exports io.ballerina.stdlib.persist.compiler.model;
+    exports io.ballerina.stdlib.persist.compiler.codeaction;
 }